Add Output Templates section in README.md
Added a section for output templates so that users of the application have an easy reference as opposed to searching within the code. 

I also updated the Output Templates subsection in the downloads section to mention both .ts and .mp4 files since the default changes in version crunchy-cli v3.0.0-dev.9 of the binaries.